requirements.txt
gym>=0.17.0
hyperopt
igraph==0.1.11
matplotlib
mushroom-rl
numpy>=1.18.1
pygame
python-igraph
scikit-learn>=0.22.1
scipy>=1.4.1
tensorflow>=2.1.0
torch>=1.3.1
tqdm
plotly
seaborn
git+https://github.com/eleurent/rl-agents
git+https://github.com/eleurent/gym-minigrid
mpi4py
pandas
statsmodels
xgboost
pytest==5.4.3
# required in opt_strategy_basic
cvxpy==1.0.3

# New optimizer
git+https://github.com/ARM-software/mango

# Web logging
neptune-client